requirements.txt
flask
mysqlclient
flask_sqlalchemy
requests
flask_login